Added validations + refactored the project a little bit

+++ FinalChallengeSA.Application/Commands/Customers/DeleteCustomer/DeleteCustomerCommandHandler.cs
@@ -18,7 +18,13 @@ namespace FinalChallengeSA.Application.Commands.Customers.DeleteCustomer
             DeleteCustomerCommand command,
             CancellationToken cancellationToken)
         {
-            var _ = await _repository.GetByIdAsync(command.Id, cancellationToken) ?? throw new NotFoundException($"Customer '{command.Id}' not found.");
+            var _ = await _repository.GetByIdAsync(command.Id, cancellationToken) ?? throw new NotFoundException($"Cliente com id '{command.Id}' não encontrado.");
+            var hasOrders = await _repository.HasOrdersAsync(command.Id, cancellationToken);
+            if (hasOrders)
+            {
+                throw new ValidationException($"Não é possível deletar o Cliente '{command.Id}' pois ele possui pedidos criados.");
+            }
+
             await _repository.DeleteAsync(command.Id, cancellationToken);
         }
     }
